requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.10
pyTelegramBotAPI==3.7.4
PyYAML==5.3.1
requests==2.25.0
urllib3==1.26.2
